+++ packages/indexer/prisma/migrations/20260421115835_add_timestamp/migration.sql
@@ -4,7 +4,6 @@
   - Added the required column `createdAt` to the `Batch` table without a default value. This is not possible if the table is not empty.
   - Added the required column `createdAt` to the `Block` table without a default value. This is not possible if the table is not empty.
   - Added the required column `createdAt` to the `Settlement` table without a default value. This is not possible if the table is not empty.
-  - Added the required column `createdAt` to the `Transaction` table without a default value. This is not possible if the table is not empty.
 
 */
 -- AlterTable
@@ -15,6 +14,3 @@ ALTER TABLE "Block" ADD COLUMN     "createdAt" TIMESTAMP(3) NOT NULL;
 
 -- AlterTable
 ALTER TABLE "Settlement" ADD COLUMN     "createdAt" TIMESTAMP(3) NOT NULL;
-
--- AlterTable
-ALTER TABLE "Transaction" ADD COLUMN     "createdAt" TIMESTAMP(3) NOT NULL;