+++ packages/indexer/src/tasks/IndexBatchTask.ts
@@ -46,11 +46,7 @@ export class IndexBatchTask
       },
       fromJSON: (parameter: string) => {
         // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-        const parsed = JSON.parse(parameter) as Batch;
-        return {
-          ...parsed,
-          createdAt: new Date(parsed.createdAt),
-        };
+        return JSON.parse(parameter) as Batch;
       },
     };
   }