+++ packages/sequencer/src/settlement/interactions/bridging/BridgingSettlementInteraction.ts
@@ -121,7 +121,7 @@ export class BridgingSettlementInteraction implements SettleInteraction {
       batches: [batch.height],
       promisedMessagesHash: latestSequenceStateHash.toString(),
       transactionHash,
-      createdAt: new Date(),
+      createdAt: Date.now(),
     };
   }
 }