+++ packages/sequencer/src/settlement/interactions/vanilla/VanillaSettlementInteraction.ts
@@ -113,7 +113,7 @@ export class VanillaSettlementInteraction implements SettleInteraction {
       batches: [batch.height],
       promisedMessagesHash: latestSequenceStateHash.toString(),
       transactionHash,
-      createdAt: new Date(),
+      createdAt: Date.now(),
     };
   }
 }