+++ packages/sequencer/src/storage/model/Batch.ts
@@ -5,7 +5,7 @@ export interface Batch {
   proof: JsonProof;
   blockHashes: string[];
   height: number;
-  createdAt: Date;
+  createdAt: number;
 }
 
 export interface SettleableBatch extends Batch {