+++ packages/sequencer/src/storage/model/Settlement.ts
@@ -2,5 +2,5 @@ export interface Settlement {
   transactionHash: string;
   promisedMessagesHash: string;
   batches: number[];
-  createdAt: Date;
+  createdAt: number;
 }